MODIFY: modified useful_functions to make them more basic and modular. Fixed luminance table
