--- LBTjpeg.py
@@ -1,318 +1,0 @@
-from typing import Tuple, NamedTuple, Optional
-import numpy as np
-import matplotlib.pyplot as plt
-from .laplacian_pyramid import quant1, quant2
-from .dct import dct_ii, colxfm, regroup
-from .bitword import bitword
-from .lbt import pot_ii
-from .familiarisation import plot_image
-from visstructure import visualerror
-
-
-def jpegencLBT(X: np.ndarray, qstep: float, s = None, N: int = 8, M: int = 8,
-        opthuff: bool = False, dcbits: int = 8, log: bool = True
-        ) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
-    '''
-    Encodes the image in X to generate a variable length bit stream.
-
-    Parameters:
-        X: the input greyscale image
-        qstep: the quantisation step to use in encoding
-        N: the width of the DCT block (defaults to 8)
-        M: the width of each block to be coded (defaults to N). Must be an
-            integer multiple of N - if it is larger, individual blocks are
-            regrouped.
-        opthuff: if true, the Huffman table is optimised based on the data in X
-        dcbits: the number of bits to use to encode the DC coefficients
-            of the DCT.
-
-    Returns:
-        vlc: variable length output codes, where ``vlc[:,0]`` are the codes and
-            ``vlc[:,1]`` the number of corresponding valid bits, so that
-            ``sum(vlc[:,1])`` gives the total number of bits in the image
-        hufftab: optional outputs containing the Huffman encoding
-            used in compression when `opthuff` is ``True``.
-    '''
-
-    if M % N != 0:
-        raise ValueError('M must be an integer multiple of N!')
-
-    # DCT on input image X.
-    if log:
-        print('Forward {} x {} DCT'.format(N, N))
-    C8 = dct_ii(N)
-    if s == None:
-        Xp = X
-    else:
-        N = C8.shape[0]
-        Pf, Pr = pot_ii(N, s)
-        t = np.s_[N//2:-N//2]
-        Xp = X.copy()
-        Xp[t, :] = colxfm(Xp[t, :], Pf)
-        Xp[:, t] = colxfm(Xp[:, t].T, Pf).T
-    Y = colxfm(colxfm(Xp, C8).T, C8).T
-
-    # Quantise to integers.
-    if log:
-        print('Quantising to step size of {}'.format(qstep))
-    Yq = quant1(Y, qstep, qstep).astype('int')
-
-    # Generate zig-zag scan of AC coefs.
-    scan = diagscan(M)
-
-    # On the first pass use default huffman tables.
-    if log:
-        print('Generating huffcode and ehuf using default tables')
-    dhufftab = huffdflt(1)  # Default tables.
-    huffcode, ehuf = huffgen(dhufftab)
-
-    # Generate run/ampl values and code them into vlc(:,1:2).
-    # Also generate a histogram of code symbols.
-    if log:
-        print('Coding rows')
-    sy = Yq.shape
-    huffhist = np.zeros(16 ** 2)
-    vlc = []
-    for r in range(0, sy[0], M):
-        for c in range(0, sy[1], M):
-            yq = Yq[r:r+M,c:c+M]
-            # Possibly regroup
-            if M > N:
-                yq = regroup(yq, N)
-            yqflat = yq.flatten('F')
-            # Encode DC coefficient first
-            dccoef = yqflat[0] + 2 ** (dcbits-1)
-            if dccoef not in range(2**dcbits):
-                raise ValueError(
-                    'DC coefficients too large for desired number of bits')
-            vlc.append(np.array([[dccoef, dcbits]]))
-            # Encode the other AC coefficients in scan order
-            # huffenc() also updates huffhist.
-            ra1 = runampl(yqflat[scan])
-            vlc.append(huffenc(huffhist, ra1, ehuf))
-    # (0, 2) array makes this work even if `vlc == []`
-    vlc = np.concatenate([np.zeros((0, 2), dtype=np.intp)] + vlc)
-
-    # Return here if the default tables are sufficient, otherwise repeat the
-    # encoding process using the custom designed huffman tables.
-    if not opthuff:
-        if log:
-            print('Bits for coded image = {}'.format(sum(vlc[:, 1])))
-        return vlc, dhufftab
-
-    # Design custom huffman tables.
-    if log:
-        print('Generating huffcode and ehuf using custom tables')
-    dhufftab = huffdes(huffhist)
-    huffcode, ehuf = huffgen(dhufftab)
-
-    # Generate run/ampl values and code them into vlc(:,1:2).
-    # Also generate a histogram of code symbols.
-    if log:
-        print('Coding rows (second pass)')
-    huffhist = np.zeros(16 ** 2)
-    vlc = []
-    for r in range(0, sy[0], M):
-        for c in range(0, sy[1], M):
-            yq = Yq[r:r+M, c:c+M]
-            # Possibly regroup
-            if M > N:
-                yq = regroup(yq, N)
-            yqflat = yq.flatten('F')
-            # Encode DC coefficient first
-            dccoef = yqflat[0] + 2 ** (dcbits-1)
-            vlc.append(np.array([[dccoef, dcbits]]))
-            # Encode the other AC coefficients in scan order
-            # huffenc() also updates huffhist.
-            ra1 = runampl(yqflat[scan])
-            vlc.append(huffenc(huffhist, ra1, ehuf))
-    # (0, 2) array makes this work even if `vlc == []`
-    vlc = np.concatenate([np.zeros((0, 2), dtype=np.intp)] + vlc)
-
-    if log:
-        print('Bits for coded image = {}'.format(sum(vlc[:, 1])))
-        print('Bits for huffman table = {}'.format(
-            (16 + max(dhufftab.huffval.shape))*8))
-
-    return vlc, dhufftab
-def jpegdecLBT(vlc: np.ndarray, qstep: float, s = None, N: int = 8, M: int = 8,
-        hufftab: Optional[HuffmanTable] = None,
-        dcbits: int = 8, W: int = 256, H: int = 256, log: bool = True
-        ) -> np.ndarray:
-    '''
-    Decodes a (simplified) JPEG bit stream to an image
-
-    Parameters:
-
-        vlc: variable length output code from jpegenc
-        qstep: quantisation step to use in decoding
-        N: width of the DCT block (defaults to 8)
-        M: width of each block to be coded (defaults to N). Must be an
-            integer multiple of N - if it is larger, individual blocks are
-            regrouped.
-        hufftab: if supplied, these will be used in Huffman decoding
-            of the data, otherwise default tables are used
-        dcbits: the number of bits to use to decode the DC coefficients
-            of the DCT
-        W, H: the size of the image (defaults to 256 x 256)
-
-    Returns:
-
-        Z: the output greyscale image
-    '''
-
-    opthuff = (hufftab is not None)
-    if M % N != 0:
-        raise ValueError('M must be an integer multiple of N!')
-
-    # Set up standard scan sequence
-    scan = diagscan(M)
-
-    if opthuff:
-        if len(hufftab.bits.shape) != 1:
-            raise ValueError('bits.shape must be (len(bits),)')
-        if log:
-            print('Generating huffcode and ehuf using custom tables')
-    else:
-        if log:
-            print('Generating huffcode and ehuf using default tables')
-        hufftab = huffdflt(1)
-    # Define starting addresses of each new code length in huffcode.
-    # 0-based indexing instead of 1
-    huffstart = np.cumsum(np.block([0, hufftab.bits[:15]]))
-    # Set up huffman coding arrays.
-    huffcode, ehuf = huffgen(hufftab)
-
-    # Define array of powers of 2 from 1 to 2^16.
-    k = 2 ** np.arange(17)
-
-    # For each block in the image:
-
-    # Decode the dc coef (a fixed-length word)
-    # Look for any 15/0 code words.
-    # Choose alternate code words to be decoded (excluding 15/0 ones).
-    # and mark these with vector t until the next 0/0 EOB code is found.
-    # Decode all the t huffman codes, and the t+1 amplitude codes.
-
-    eob = ehuf[0]
-    run16 = ehuf[15 * 16]
-    i = 0
-    Zq = np.zeros((H, W))
-
-    if log:
-        print('Decoding rows')
-    for r in range(0, H, M):
-        for c in range(0, W, M):
-            yq = np.zeros(M**2)
-
-            # Decode DC coef - assume no of bits is correctly given in vlc table.
-            cf = 0
-            if vlc[i, 1] != dcbits:
-                raise ValueError(
-                    'The bits for the DC coefficient does not agree with vlc table')
-            yq[cf] = vlc[i, 0] - 2 ** (dcbits-1)
-            i += 1
-
-            # Loop for each non-zero AC coef.
-            while np.any(vlc[i] != eob):
-                run = 0
-
-                # Decode any runs of 16 zeros first.
-                while np.all(vlc[i] == run16):
-                    run += 16
-                    i += 1
-
-                # Decode run and size (in bits) of AC coef.
-                start = huffstart[vlc[i, 1] - 1]
-                res = hufftab.huffval[start + vlc[i, 0] - huffcode[start]]
-                run += res // 16
-                cf += run + 1
-                si = res % 16
-                i += 1
-
-                # Decode amplitude of AC coef.
-                if vlc[i, 1] != si:
-                    raise ValueError(
-                        'Problem with decoding .. you might be using the wrong hufftab table')
-                ampl = vlc[i, 0]
-
-                # Adjust ampl for negative coef (i.e. MSB = 0).
-                thr = k[si - 1]
-                yq[scan[cf-1]] = ampl - (ampl < thr) * (2 * thr - 1)
-
-                i += 1
-
-            # End-of-block detected, save block.
-            i += 1
-
-            yq = yq.reshape((M, M)).T
-
-            # Possibly regroup yq
-            if M > N:
-                yq = regroup(yq, M//N)
-            Zq[r:r+M, c:c+M] = yq
-
-    if log:
-        print('Inverse quantising to step size of {}'.format(qstep))
-
-    Zi = quant2(Zq, qstep, qstep)
-
-    if log:
-        print('Inverse {} x {} DCT\n'.format(N, N))
-    C8 = dct_ii(N)
-    Z = colxfm(colxfm(Zi.T, C8.T).T, C8.T)
-    if s == None:
-        return Z
-    else:
-        N = C8.shape[0]
-        Pf, Pr = pot_ii(N, s)
-        t = np.s_[N//2:-N//2]
-        Zp = Z.copy()
-        Zp[:, t] = colxfm(Zp[:, t].T, Pr.T).T
-        Zp[t, :] = colxfm(Zp[t, :], Pr.T)
-        return Zp
-    return Z
-def eightLBTjpegtune(X):
-    X = X-128.0
-    SSIMscore = 0
-    for k in range(120, 150, 2):
-        step_size = 16
-        calc_bits = 100000
-        k = k/100
-        while calc_bits > 40960:
-            vlctemp, _ = jpegencLBT(X, step_size, s=k)
-            calc_bits = vlctemp[:,1].sum()
-            step_size += 1
-        vlc, _ = jpegencLBT(X, step_size, k)
-        Z = jpegdecLBT(vlc, step_size, k)
-        rmserr = np.std(Z - X)
-        scoretemp = visualerror(X, Z)
-        print(scoretemp)
-        if scoretemp > SSIMscore:
-            SSIMscore = scoretemp
-            Zfinal = Z
-    return Zfinal, rmserr, calc_bits, SSIMscore
-def eightDCTjpeg(X):
-    X = X- 128.0
-    step_size = 16
-    calc_bits = 100000
-    while calc_bits > 40960:
-        vlctemp, _ = jpegenc(X, step_size)
-        calc_bits = vlctemp[:,1].sum()
-        step_size += 1
-    vlc, _ = jpegenc(X, step_size)
-    Z = jpegdec(vlc, step_size)
-    rmserr = np.std(Z - X)
-    Verr = visualerror(X, Z)
-    return Z, rmserr, calc_bits, Verr
-def compress(X):
-    ZLBT, eLBT, bLBT, scoreLBT = eightLBTjpegtune(X)
-    ZDCT, eDCT, bDCT, scoreDCT = eightDCTjpeg(X)
-    if scoreDCT > scoreLBT:
-        print(f"decoded with an rms error {eDCT} using {bDCT} bits and an SSIM score of {scoreDCT} (using DCT)")
-        fig, ax = plt.subplots()
-        plot_image(ZDCT, ax=ax)
-    else:
-        print(f"decoded with an rms error {eLBT} using {bLBT} bits and an SSIM score of {scoreLBT} (using LBT)")
-        fig, ax = plt.subplots()
-        plot_image(ZLBT, ax=ax)
