--- test.py
@@ -1,4 +1,0 @@
-for i in range(32):
-    for j in range(32):
-        print(i % 8, j % 8)
-        print(i, j)
