+++ testing_luminance_table_dct.py
@@ -17,16 +17,16 @@ X = X_pre_zero_mean - 128.0
 # JPEG quantisation luminance table (section K.1)
 
 step_table = np.array([
-    [16, 11, 10, 16, 124, 140, 151, 161],
-    [12, 12, 14, 19, 126, 158, 160, 155],
-    [14, 13, 16, 24, 140, 157, 169, 156],
-    [14, 17, 22, 29, 151, 187, 180, 162],
-    [18, 22, 37, 56, 168, 109, 103, 177],
-    [24, 35, 55, 64, 181, 104, 113, 192],
+    [16, 11, 10, 16, 24, 40, 51, 61],
+    [12, 12, 14, 19, 26, 58, 60, 55],
+    [14, 13, 16, 24, 40, 57, 69, 56],
+    [14, 17, 22, 29, 51, 87, 80, 62],
+    [18, 22, 37, 56, 68, 109, 103, 77],
+    [24, 35, 55, 64, 81, 104, 113, 92],
     [49, 64, 78, 87, 103, 121, 120, 101],
-    [72, 92, 95, 98, 112, 100, 103, 199]
+    [72, 92, 95, 98, 112, 100, 103, 99]
 ], dtype=np.float64)
-step_table *= 1.5
+step_table *= 1.0
 """
 ones_array = np.ones((8, 8), dtype=np.float64)
 step_table = ones_array*23
@@ -36,8 +36,9 @@ C8 = dct_ii(8)
 qY = gen_Y_quant_dct_jpeg(X, step_table, C8, supp_comp_num=0)
 Z = colxfm(colxfm(qY.T, C8.T).T, C8.T)
 X_quant_ent = entropy(quantise(X, 17))
-qY_ent = entropy(qY)
-comp_ratio = X_quant_ent / qY_ent
+Yr = regroup(qY, 8)
+Yr_ent = dctbpp(Yr, 8)
+comp_ratio = X_quant_ent / Yr_ent
 print(comp_ratio)
 print(np.std(X - Z))
 
@@ -46,5 +47,6 @@ plot_image(Z, ax=ax)
 plt.show()
 
 step_ratio = find_step_ratio_equal_rms_dct_jpeg(X, step_table, C8)
+print(step_ratio)
 print(np.std(quantise(X, 17) - X))
 print(compute_err_dct_jpeg(X, step_ratio, step_table, C8))
